requirements.txt
tensorflow>=2.8
numpy
pillow
faiss-cpu
scikit-learn
streamlit
opencv-python
pandas
tqdm
